refactor: reduce RemoteSysdigImageScanner to a thin orchestrator
Promote ScannerPaths to a top-level class and move the CLI process execution
and exit-code handling into ScannerExecutor. RemoteSysdigImageScanner now only
wires the collaborators and coordinates the flow: create workspace -> retrieve
binary -> execute -> deserialize. Public performScan() signature and the
SysdigImageScanner caller are unchanged.
